+++ api/core/utils/responses.py
@@ -1,19 +1,32 @@
 from typing import Any, Optional
 
+from fastapi.responses import JSONResponse
 from pydantic import BaseModel
 
 
-class MyResponse(BaseModel):
-    """
-    Base response model for API responses.
-    responde status from fastapi.status
-    - https://fastapi.tiangolo.com/advanced/status-codes/
-    - https://fastapi.tiangolo.com/tutorial/response-status-code/
+class ExchangeAuthError(Exception):
+    pass
 
-    """
 
+class AppNotFoundError(Exception):
+    pass
+
+
+class MyResponse(JSONResponse):
+    def __init__(
+        self,
+        status_code: int,
+        message: Optional[str] = None,
+        data: Optional[Any] = None,
+    ):
+        content = {
+            "message": message,
+            "data": data,
+        }
+        super().__init__(status_code=status_code, content=content)
+
+
+class MyResponseModel(BaseModel):
     message: Optional[str] = None
     data: Optional[Any] = None
-
-    def __init__(self, message: Optional[str] = None, data: Optional[dict] = None):
-        super().__init__(message=message, data=data)
+    # add `status: int` here if you want it in the body